pinctrl: msm8976: Fix QDSD and SDC3 pin declarations
The QDSD pins were WRONG!!!! That's all because the SDC3 pins
were also ... wrong.

SDC3 is on "normal" GPIO array, so not the "extra" ones.
Remove the sdc3 declaration from QDSD pingroup, move it to
standard pingroup and assign the right GPIOs.
Also shift down the QDSD pins in order to finally get them
(hopefully) right.
Signed-off-by: Mohd Faraz <androiabledroid@gmail.com>